CARL ResourceManager: Move `Persistable` implementations into `resource::types`.
