+++ opendut-carl/src/resource/types/persistable.rs
@@ -1,0 +1,74 @@
+use crate::resource::manager::persistence::persistable::{Persistable, StorageKind};
+use opendut_model::cluster::{ClusterDescriptor, ClusterDeployment};
+use opendut_model::peer::configuration::{PeerConfiguration, EdgePeerConfigurationState};
+use opendut_model::peer::state::PeerConnectionState;
+use opendut_model::peer::PeerDescriptor;
+
+#[cfg(feature = "viper")]
+use opendut_model::viper::{ViperRunDeployment, ViperTestRunDescriptor, ViperSourceDescriptor};
+
+
+impl Persistable for ClusterDeployment {
+    type Proto = opendut_model::proto::cluster::ClusterDeployment;
+    const TABLE: &'static str = "cluster_deployment";
+    const STORAGE: StorageKind = StorageKind::Persistent;
+}
+
+impl Persistable for ClusterDescriptor {
+    type Proto = opendut_model::proto::cluster::ClusterDescriptor;
+    const TABLE: &'static str = "cluster_descriptor";
+    const STORAGE: StorageKind = StorageKind::Persistent;
+}
+
+impl Persistable for PeerConfiguration {
+    type Proto = opendut_model::proto::peer::configuration::api::PeerConfiguration;
+    const TABLE: &'static str = "peer_configuration";
+    /// Not persisted at the moment. A restart will cause a reconfiguration of all peers.
+    /// The `assign_cluster()` method in the `ClusterManager` evaluates the current peer descriptors of the cluster and sends new peer configurations to the peers.
+    /// It is called by the `ClusterManager` when a cluster deployment is created or when all peers of a cluster deployment are available.
+    /// -> subscription triggers following chain: `schedule_redeploying_clusters_when_all_peers_become_available()` -> `rollout_all_clusters_containing_newly_available_peer()` -> `rollout_cluster_if_all_peers_available()`
+    /// This is done to ensure that the cluster deployment is always up-to-date and that all peers are configured correctly.
+    /// It is intended to be stored in a persistent storage once the peer configuration is stable to avoid migration issues.
+    ///
+    /// If you want to persist this, make sure the hashing algorithm in the implementations of `ParameterValue` is stable across Rust releases.
+    const STORAGE: StorageKind = StorageKind::Volatile;
+}
+
+impl Persistable for PeerConnectionState {
+    type Proto = opendut_model::proto::peer::PeerConnectionState;
+    const TABLE: &'static str = "peer_connection_state";
+    const STORAGE: StorageKind = StorageKind::Volatile;
+}
+
+impl Persistable for PeerDescriptor {
+    type Proto = opendut_model::proto::peer::PeerDescriptor;
+    const TABLE: &'static str = "peer_descriptor";
+    const STORAGE: StorageKind = StorageKind::Persistent;
+}
+
+impl Persistable for EdgePeerConfigurationState {
+    type Proto = opendut_model::proto::peer::configuration::api::EdgePeerConfigurationState;
+    const TABLE: &'static str = "peer_configuration_state";
+    const STORAGE: StorageKind = StorageKind::Volatile;
+}
+
+#[cfg(feature = "viper")]
+impl Persistable for ViperSourceDescriptor {
+    type Proto = opendut_model::proto::viper::ViperSourceDescriptor;
+    const TABLE: &'static str = "viper_source_descriptor";
+    const STORAGE: StorageKind = StorageKind::Persistent;
+}
+
+#[cfg(feature = "viper")]
+impl Persistable for ViperTestRunDescriptor {
+    type Proto = opendut_model::proto::viper::ViperTestRunDescriptor;
+    const TABLE: &'static str = "viper_test_run_descriptor";
+    const STORAGE: StorageKind = StorageKind::Persistent;
+}
+
+#[cfg(feature = "viper")]
+impl Persistable for ViperRunDeployment {
+    type Proto = opendut_model::proto::viper::ViperRunDeployment;
+    const TABLE: &'static str = "viper_run_deployment";
+    const STORAGE: StorageKind = StorageKind::Persistent;
+}
